requirements.txt
django
djangorestframework
psycopg2

whitenoise